// FamilyClub.React/famiy.club.react/lib/api/generated/models/ComplaintsReadDto.ts
export interface ComplaintsReadDto {
    /**
     * 
     * @type {number}
     * @memberof ComplaintsReadDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ComplaintsReadDto
     */
    complaintText?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComplaintsReadDto
     */
    complaintType?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ComplaintsReadDto
     */
    isResolved?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ComplaintsReadDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ComplaintsReadDto
     */
    resolvedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ComplaintsReadDto
     */
    clubMemberId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComplaintsReadDto
     */
    resolutionNotes?: string | null;
    /**
     * 
     * @type {Array<ComplaintImageDto>}
     * @memberof ComplaintsReadDto
     */
    images?: Array<ComplaintImageDto> | null;
}